在安装脚本中添加 skipd 版本判断语句
* 若非 skipd v3 版本，则停止安装

+++ cfddns/install.sh
@@ -1,5 +1,10 @@
 #!/bin/sh
 
+if [[ `dbus version|awk -F " " '{print $NF}'` !=  "3.0.0" ]]; then
+    echo "dbus version (`dbus version|awk -F " " '{print $NF}'`) is not supported now, please contact the developer!"
+    exit
+fi
+
 cp -r /tmp/cfddns/res/* /koolshare/res
 cp -r /tmp/cfddns/scripts/* /koolshare/scripts
 cp -r /tmp/cfddns/webs/* /koolshare/webs
